new addon: server_monitoring
This module allows in-database logging of some statistics in order to monitor
the health of an openerp instance.

Database indicators are logged (number of rows, table size, number of reads,
number of updates...), with a cron running each week by default. This cron
needs to be activated after the module is installed.

Some process indicators are logged (cpu time, memory) together with information
about the different XMLRPC calls made to the server (user, model, method).

Two crons are provided to cleanup old logs from the database.

The logs are available through the menu Reporting -> Server Monitoring.