Remove the company on record.lifespan
The company was mandatory and prevented to use the module for models without
company_id.  Also, there is currently no use case for a different configuration
for different companies.  It simplifies the setup too.

If a use case exists for a different setup per company, it can be reintroduced
later with a optional company_id.