Rename addon according to the new term: partner_changeset
Add translations

Add test coverage for caching of changeset rules

Add unique constraint (model_id, field_id) on rules

Add rules in demo data

Put the security groups links in noupdate

Put keep the groups in a noupdate=0 section so the records in ir.model.data
will still be modifiable by other modules.

Store the source of a changeset

Differentiate rules according to their origin

Rules are applied also for manual edition

Action is required

Do not keep recordsets in ormcache

Because they would be unreadable as soon as the cursor is closed.
Instead, we keep only the id and the record is browsed for every new
environment.

Remove useless 'model_id' on changeset rules

model_id has been removed

Remove reference to the model_name on rules

Because the model has been removed earlier (dead code)

Fix issue when applying empty many2one

Do not create changesets on moved contacts

As we just created the contact with a 'copy' we don't want to have a
changeset for the initialization values

Update translations

Use a selection widget on source model

Adapt for inclusion in OCA

Rename 'Pending Changesets' to 'Changes'

It's shorter

Add screenshots

Do not create a changeset when both sides are empty

But have a different type (e.g. False and '')