[ENH] module_prototyper: Refactor the code.
The original `topo.py` reflected a part of the algorithm which is not the
essence of what we need.  Find a *closure* in the graph of records is the
actual need.  The topological sort algorithm is used only to warrant that
records are given in a way that it makes sense for generating the XML.