auditlog: add rules per field
This commit adds ability to create log records based on per field rules.
User can specify which field needs to be tracked for any CRUD operation.
This doesn't break previous configurations. If rule is set on model, it
will be applied to all fields. But if model hasn't got any particular
rule, then patching method will look for field rules.

It is not possible to specify full/fast log type per field, and it is
taken from model.