[IMP] users_ldap_populate. Make it safe to delete unknown users.
To make it safe deactivate unknown users we had to know which users have been created from
ldap. This also gives the possibility to differentiate other actions based on wether user was
created from ldap or added to the database in another way. For instance ldap users should not
be able to reset their password in Odoo, because they must do it in the LDAP server (for instance
through ldap_admin, or through a Windoes user account).