auditlog: Fix fast-logging exception on create with multiple records
When fast-logging create calls the code assumed the old-style
calling convention where only one dictionary value is passed
into the create() function. When it received a list of
dictionaries instead it blew up trying to manipulate a
single dict object instead of a list.

o Fix the code to degrade gracefully to full logging, instead.
o Add a unit test to catch this case in the future
o Fixes #1619