Store the 'server.config' fields in sparse fields
This model is used only to display the configuration to authorized
users. It's a TransientModel that dynamically creates the fields for
every odoo configuration parameter, environment variables and some
system ones. Before this change, each new parameter/variable will
create a field in database, which can lead to many issues (lot of
fields, field name too long). Storing the fields in a single JSON
field resolves these issues and is much cleaner anyway.