[FIX] Auditlog: avoid access rights errors while handling creation record
This commit fixes following case:
When record has fields that user has no access to, and user tries to
create such record, the after record created auditlog tries to read all
fields of record, but record creation fails because user cannot read one
of fields.

This fix was caught on db where full-log was enabled on Partner model,
and some user has no access to signup_token field. So he was not able to
create partners.