[10.0][FIX] Authadmin passkey doesn't validate param before eval
Without being validated, if get_param return False, it will try to
pass False to safe_eval that will raise an exception trying to
evaluate False which isn't a string.

As a result, a user typing to password of an admin while auth passkey
params aren't configured will see an internal server error... Which
isn't as "safe" as just showing bad credentials.