[REF] make menu more consistent
In normal tree view duplicate and archive can be done throught the
action menu. So we only need to show this button when this tree view
are included in a form using a o2m or m2m.
So this visibility can be activated by passing value in the context

Also make the copy consitent whatever way you do it