[FIX] letsencrypt. Use char instead of text fields for config.
res.config.settings does not allow Text fields for config_parameter fields. The solution to overcome this is generating
warnings in runbot. As varchar fields without size are to all intents and purposes the same as text fields in postgress,
and can be painlessly and automatically converted, the field type is changed.