[13.0][FIX] base_changeset
When fields view get method for form view is executed, self does not
contain any record. This is causing the validation pop-ups to never
display because the length of self is 0, and the computed value user_
can_see_changeset can't get properly evaluated.

To fix it I am proposing:

- Use method directly instead of computed field to check user rights
to the changeset, as the compute method does not evaluate correctly
in empty record.

- Do not display button box if no changesets are pending.