[13.0][REF] base_changeset: remove base model inheritance
Using the base model implies adding the changeset fields and
methods for all models in the system. This is causing errors
when the module is updated, as recorded in issue #2187.

Usually the practice is not to perform this kind of inheritance,
but to create a new Abstract Model to be used to add such
functionalities.

This MR intends to perform this modifications so that the models that
should be provided with the changeset fields and methods can simply
inherit from this new Abstract Model. To do so, new modules can be
developed for each desired behaviour (for example purchase_changeset,
sale_changeset...).