fix(database_cleanup): endless loop
I had a case where database_cleanup looped endlessly because of
referential integrity. DROP TABLE was failing because some tables had
foreign keys on tables which were not in the list of tables to delete.
Dropping all contraints linked to the table, even if they are on table which won't be dropped seems to solve the issue.