[FIX] base_changeset: fix caching
ormcache is used to minimize the number of times that the active changeset
rules are quered, but it was still missing in one important place.

In this module's tests, this reduces the number of calls to
`models_to_track_changeset` from 354 to 135 times, and the number of test
queries is reduced from 4069 to 3810.