[REF] base_sequence_default: bump sequence only on creation
Before this patch, sequence was bumped when loading defaults. That happened even before creating the record. That rendered no-gap sequences useless.

Now, sequences are executed only on record creation, when the user didn't give a default value to the sequenced fields.

To support the use case when you want a sequence on required fields, you can set a value of `-` on those fields on creation. Then, the sequence will also kick in.

@moduon MT-2715