[16.0-fix-multi-company] [FIX] Multi-Company issue with M2M field
In a Multi-Company environment with the auditlog rule on product.template enabled, the taxes_id field gets always overwritten with the taxes chosen in the current company.
The cause of this is that the old_values are read with sudo, which has access to not only the current company.
When doing the actual write of the field, this cached value is taken into account, which removes them from the M2M relation.
Before doing the actual write, the cache of the recordset(s) need to be cleared again.