[IMP] tracking_manager: Fix trackable conditions
The condition on stored field does not allow to track changes
on property fields as they are not stored on the model itself,
although Odoo allows to track such fields.

Use a blacklist of mixins to make sure all the fields from these
models are blacklisted and use a function to allow custom redefinition.