[FIX] session_db: Fix creation of postgres_uri in tests
with the parameter "login" in `_make_postgres_uri()` it becomes impossible to log in with the configured database, as it has to have the name "user". This is because of passing the `connection_info` dict as keywords parameter into the function. But because `connection_info_for` called in `setUp()` reads the config file of odoo and there the database user is `db_user` the keyword parameter is discarded and `login` is not set. The `db_`-prefix is removed. So user and password are not applied to the uri/dsn and thus login to the database is not possible, if the database wants those parameters.

When I applied this change in our internal CI of solute.de, everything worked fine afterwards.

This has to be ported to 17.0 and 18.0 as well.