[FIX] test.yml: prevent infinite loop of makepot commits
the result will be the same in each of those jobs because the pot files will be
generated for all the addons at once.

If there are more than one test run with different sets of modules, and in one of
those modules a field is added to the base model, the makepot jobs will have different
outcomes and will keep on reverting each other's changes upon subsequent test runs.