[FIX] test.yml: prevent infinite loop of makepot commits
Don't run makepot using both Odoo and OCB. This is redundant, but also, because
of changes between both versions, the POTs may have trivial differences. The
commit of a makepot run will again trigger CI to run makepot with both versions,
leading to an infinite loop of competing commits trying to revert each other's
trivial differences.

In the current case, the sms module is set to auto_install in Odoo but not in
OCB. This causes the field `message_has_sms_error` from sms' mail.thread override
to show up in the Odoo generated POT files, but not in the OCB ones.