Remove pygount hardcoded version and improve compatibility
pygount 1.4.0 is now very old and current code does not work on recent versions.
It is however easy to improve compatibility accross pygount versions using the
code_count property instead of the code one, which was removed in 2.0.0.
That way, pygount 1.4.0 stays supported but also modern version up to at least
 3.1.0.