[FIX] auditlog: swap out tocompute in ThrowAwayCache as well
Steps to Reproduce:

Install auditlog and purchase_tier_validation. Create an auditlog rule on
the Tier Exception model. Create a new tier exception record for the purchase
order model without any fields. In an export, check the value for the stored
computed field `model_name`.

Desired outcome:

The value for `model_name` is 'purchase.order'.

Actual outcome:

There is no value for the stored computed field `model_name`. When trying to
edit the `Fields` model, which depends on a valid value for this field, an error
occurs saying 'Compute method failed to assign
tier.validation.exception(1,).valid_model_field_ids'.

Analysis:

When `read` is executed under the scope of the new `ThrowAwayCache` decorator
(#3374), recomputes may be performed and dequeued from
`self.env.cache.transaction.tocompute`. To make sure the recomputed values still
end up in the non-volatile as well, we need to preserve the tocompute attribute
from the original transaction.